@@ -3,27 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "moment": "^2.26.0",
-    "react": "^17.0.2",
-    "react-big-calendar": "^0.38.0",
-    "react-datetime-picker": "^3.0.2",
-    "react-dom": "^17.0.2",
-    "react-modal": "^3.11.2",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "sweetalert2": "^11.1.9"
+    "@reduxjs/toolkit": "^2.6.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "moment": "^2.30.1",
+    "react": "^19.1.0",
+    "react-big-calendar": "^1.18.0",
+    "react-datetime-picker": "^6.0.1",
+    "react-detect-offline": "^2.4.5",
+    "react-dom": "^19.1.0",
+    "react-modal": "^3.16.3",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^6.29.0",
+    "react-scripts": "5.0.1",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0",
+    "sweetalert2": "^11.17.2"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "react-scripts build && yarn PWA",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "PWA": "workbox injectManifest"
   },
   "eslintConfig": {
     "extends": "react-app"
