@@ -1,56 +1,51 @@
 import React, { useEffect } from 'react';
 import {
     BrowserRouter as Router,
-    Switch,
-    Redirect
-  } from 'react-router-dom';
-  
+    Routes,
+    Route,
+    Navigate
+} from 'react-router-dom';
+
 import { useDispatch, useSelector } from 'react-redux';
 
 import { LoginScreen } from '../components/auth/LoginScreen';
 import { CalendarScreen } from '../components/calendar/CalendarScreen';
 import { startChecking } from '../actions/auth';
-import { PublicRoute } from './PublicRoute';
-import { PrivateRoute } from './PrivateRoute';
 
 export const AppRouter = () => {
-
     const dispatch = useDispatch();
-    const { checking, uid } = useSelector( state => state.auth);
+    const { checking, uid } = useSelector(state => state.auth);
 
     useEffect(() => {
-        
-        dispatch( startChecking() );
+        dispatch(startChecking());
+    }, [dispatch]);
 
-    }, [dispatch])
-
-    if ( checking ) {
-        return (<h5>Espere...</h5>);
+    if (checking) {
+        return <h5>Espere...</h5>;
     }
-
 
     return (
         <Router>
-            <div>
-                <Switch>
+            <Routes>
+                {/* Ruta pública */}
+                <Route
+                    path="/login"
+                    element={
+                        uid ? <Navigate to="/" /> : <LoginScreen />
+                    }
+                />
 
-                    <PublicRoute 
-                        exact 
-                        path="/login" 
-                        component={ LoginScreen }
-                        isAuthenticated={ !!uid }
-                    />
+                {/* Ruta privada */}
+                <Route
+                    path="/"
+                    element={
+                        uid ? <CalendarScreen /> : <Navigate to="/login" />
+                    }
+                />
 
-                    <PrivateRoute 
-                        exact 
-                        path="/" 
-                        component={ CalendarScreen } 
-                        isAuthenticated={ !!uid }
-                    />
-
-                    <Redirect to="/" />   
-                </Switch>
-            </div>
+                {/* Redirección por defecto */}
+                <Route path="*" element={<Navigate to="/" />} />
+            </Routes>
         </Router>
-    )
-}
+    );
+};
