@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Route, Redirect } from 'react-router-dom';
+import { Route, Navigate } from 'react-router-dom';
 
 
 export const PublicRoute = ({
@@ -12,11 +12,8 @@ export const PublicRoute = ({
 
     return (
         <Route { ...rest }
-            component={ (props) => (
-                ( isAuthenticated )
-                    ? ( <Redirect to="/" /> )
-                    : ( <Component { ...props } /> )
-            )}
+            element={isAuthenticated ? <Navigate to="/" /> : <Component/>}
+            
         
         />
     )
